+++ wsidicom/wsidicom.py
@@ -815,6 +815,14 @@ class WsiDicom:
         if self._source_owned:
             self._source.close()
 
+    def clear_cache(self):
+        """Clear cache of encoded and decoded tiles."""
+        self._source.clear_cache()
+
+    def resize_cache(self, size: int):
+        """Resize cache of encoded and decoded tiles."""
+        self._source.resize_cache(size)
+
     def _validate_collection(self) -> SlideUids:
         """
         Check that no files or instance in collection is duplicate, and, if